Fileferry/fileferry-agent/requirements.txt
# FileFerry Agent - Requirements

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Slack Integration
slack-sdk==3.27.0
slack-bolt==1.18.1

# AWS Integration
boto3==1.29.7
botocore==1.32.7
aioboto3==12.0.0

# Azure Integration
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-core==1.29.5

# HTTP/Async
aiohttp==3.9.1
httpx==0.25.2

# Monitoring - Datadog
datadog==0.48.0
ddtrace==2.3.0

# File Transfer
paramiko==3.4.0
pysftp==0.2.9

# Progress Tracking & Retry Logic
tqdm==4.66.1
tenacity==8.2.3

# ServiceNow & HTTP
requests==2.31.0

# Configuration & Logging
python-dotenv==1.0.0
pyyaml==6.0.1

# Date/Time utilities
python-dateutil==2.8.2

# Security
cryptography==41.0.7

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==5.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Legacy (kept for compatibility)
flask==3.0.0
docker==7.0.0
awscli==1.31.7